apply biome format

+++ src/lib/settings.ts
@@ -1,10 +1,10 @@
-import { and, eq } from "drizzle-orm";
-import { db } from "~/server/db";
-import { settings, userSettings } from "~/server/db/schema";
+import { and, eq } from 'drizzle-orm'
+import { db } from '~/server/db'
+import { settings, userSettings } from '~/server/db/schema'
 
 export async function getSetting<T>(key: string): Promise<T | null> {
     const setting = await db.query.settings.findFirst({
-        where: eq(settings.key, key)
+        where: eq(settings.key, key),
     })
 
     return (setting?.value as T) ?? null
@@ -12,14 +12,14 @@ export async function getSetting<T>(key: string): Promise<T | null> {
 
 export async function setSetting<T>(key: string, value: T) {
     try {
-        await db.insert(settings)
-            .values({
-                key,
-                value,
-            })
+        await db.insert(settings).values({
+            key,
+            value,
+        })
     } catch (error) {
         // update
-        await db.update(settings)
+        await db
+            .update(settings)
             .set({
                 value: value,
             })
@@ -28,14 +28,13 @@ export async function setSetting<T>(key: string, value: T) {
 }
 
 export async function deleteSetting(key: string) {
-    const r = await db.delete(settings)
-        .where(eq(settings.key, key))
+    const r = await db.delete(settings).where(eq(settings.key, key))
     return r.rowCount == 1
 }
 
 export async function getUserSetting<T>(key: string, userId: string): Promise<T | null> {
     const setting = await db.query.userSettings.findFirst({
-        where: and(eq(userSettings.key, key), eq(userSettings.userId, userId))
+        where: and(eq(userSettings.key, key), eq(userSettings.userId, userId)),
     })
 
     return (setting?.value as T) ?? null
@@ -43,26 +42,23 @@ export async function getUserSetting<T>(key: string, userId: string): Promise<T 
 
 export async function setUserSetting<T>(key: string, userId: string, value: T) {
     try {
-        await db.insert(userSettings)
-            .values({
-                key,
-                userId,
-                value,
-            })
+        await db.insert(userSettings).values({
+            key,
+            userId,
+            value,
+        })
     } catch (error) {
         // update
-        await db.update(userSettings)
+        await db
+            .update(userSettings)
             .set({
                 value: value,
             })
-            .where(and(eq(userSettings.key, key), eq(userSettings.userId, userId))
-            )
+            .where(and(eq(userSettings.key, key), eq(userSettings.userId, userId)))
     }
-
 }
 
 export async function deleteUserSetting(key: string, userId: string) {
-    const r = await db.delete(userSettings)
-        .where(and(eq(userSettings.key, key), eq(userSettings.userId, userId)))
+    const r = await db.delete(userSettings).where(and(eq(userSettings.key, key), eq(userSettings.userId, userId)))
     return r.rowCount == 1
 }
